requirements.txt
accelerate
bitsandbytes
datasets
evaluate
git-lfs
huggingface-hub
immutabledict
jupyter
lightning
ninja
nltk
numpy
openai
pandas
pyarrow
pyyaml
scipy
sentencepiece
tensorboard
torch
torchvision
torchaudio
transformers
wandb
